WIP commit: - first draft of the new file structure - created a bootstrapper and dependency resolver
